Initial commit on GitHub.
This is the initial commit on GitHub of
'CADEE: Computer-Aided Directed Evolution of Enzymes'.

The earlier version history was dropped, because executables,
tests and other binaries had to be removed prior to publishing.

Scripts located in the 'qscripts' sub-folder, were written by
Miha Purg. CADEE is using qscripts, to automatically analyze
runs on the fly.