Major User-Interface Overhaul
This commit is adding a number of changes to the user-interface of CADEE.

CADEE includes now a (setuptools) setup.py script to install CADEE: python setup.py install --user.
This setup is hiding most of the source code and offers the user a script named 'cadee', which is installed by setup.py.
The cadee-script is unifying access to user-accessible parts of cadee and providing a simple cli, that has the same style as Q.
Usage:
      cadee [ prep(p) | dyn(d) | analyse(a) | tool(t) ]

Example:
      cadee prep --help

The folder-structure has been arranged in the style of Q; prep; dyn; analyze.

The qscripts_config.py was adjusted, so that executables are looked up with CADEE and qscripts.cfg is automatically generated.